fam_pay_test/fam_pay_test/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
djangorestframework==3.12.4
idna==3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
netifaces==0.11.0
